Refactor SimulationPlot to more generic Plot
