+++ core/src/main/java/info/openrocket/core/simulation/DataType.java
@@ -6,4 +6,5 @@ import info.openrocket.core.util.UnitValue;
  * A type of data that can be stored in a {@link DataBranch}.
  */
 public interface DataType extends UnitValue {
+	String getName();
 }
